+++ packages/chain/mempool/interface.go
@@ -1,0 +1,30 @@
+package mempool
+
+import (
+	"time"
+
+	"github.com/iotaledger/wasp/packages/iscp"
+)
+
+type Mempool interface {
+	ReceiveRequests(reqs ...iscp.RequestData)
+	ReceiveRequest(req iscp.RequestData) bool
+	RemoveRequests(reqs ...iscp.RequestID)
+	ReadyNow(nowis ...time.Time) []iscp.RequestData
+	ReadyFromIDs(nowis time.Time, reqIDs ...iscp.RequestID) ([]iscp.RequestData, []int, bool)
+	HasRequest(id iscp.RequestID) bool
+	GetRequest(id iscp.RequestID) iscp.RequestData
+	Info() MempoolInfo
+	WaitRequestInPool(reqid iscp.RequestID, timeout ...time.Duration) bool // for testing
+	WaitInBufferEmpty(timeout ...time.Duration) bool                       // for testing
+	Close()
+}
+
+type MempoolInfo struct {
+	TotalPool      int
+	ReadyCounter   int
+	InBufCounter   int
+	OutBufCounter  int
+	InPoolCounter  int
+	OutPoolCounter int
+}
